+++ EssentialFeed/Feed Presentation/FeedPresenter.swift
@@ -1,0 +1,82 @@
+//
+//  FeedPresenter.swift
+//  EssentialFeed
+//
+//  Created by Firdavs Bagirov on 28/08/22.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public var isLoading: Bool
+}
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: FeedErrorViewModel {
+        return FeedErrorViewModel(message: nil)
+    }
+    
+    static func error(message: String) -> FeedErrorViewModel {
+        return FeedErrorViewModel(message: message)
+    }
+}
+
+public protocol FeedErrorView {
+    func display(_ viewModel: FeedErrorViewModel)
+}
+
+
+public final class FeedPresenter {
+    private let errorView: FeedErrorView
+    private let loadingView: FeedLoadingView
+    private let feedView: FeedView
+    
+    private var feedLoadError: String {
+        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Error message displayed when we can't load the image feed from the server")
+    }
+    
+    public init(errorView: FeedErrorView, loadingView: FeedLoadingView, feedView: FeedView) {
+        self.errorView = errorView
+        self.loadingView = loadingView
+        self.feedView = feedView
+    }
+    
+    public static var title: String {
+        return NSLocalizedString("FEED_VIEW_TITLE",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Title for the feed view")
+    }
+    
+    public func didStartLoadingFeed() {
+        loadingView.display(FeedLoadingViewModel(isLoading: true))
+        errorView.display(.noError)
+    }
+    
+    public func didFinishLoadingFeed(with feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoadingFeed(with error: Error) {
+        errorView.display(.error(message: feedLoadError))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
